AutoStreamAIChatbot/requirements.txt
langchain>=0.2.0
langchain-google-genai>=1.0.0
langchain-groq>=0.1.0
langchain-huggingface>=0.0.1
sentence-transformers>=2.0.0
langchain-community>=0.2.0
langchain-text-splitters>=0.2.0
langgraph>=0.1.0
chromadb>=0.5.0
pydantic>=2.0.0
python-dotenv>=1.0.0
google-generativeai>=0.7.0
streamlit>=1.35.0
